backend/requirements.txt
# BizuHub backend - dependências com versões fixas para segurança
fastapi>=0.115.0,<0.116
uvicorn[standard]>=0.32.0,<0.33
pydantic[email]>=2.0,<3
pydantic-settings>=2.0,<3
sqlalchemy>=2.0,<3
psycopg2-binary>=2.9.9,<3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0,<0.28
slowapi>=0.1.9,<0.2
langchain-groq>=0.2.0,<0.3
langchain-core>=0.3.0,<0.4
alembic>=1.13.0,<2
